# backend/src/server.rs
use crate::{
    auth::GeoEngineAuthMiddlewareLayer,
    collection_transactions::NoCollectionTransactions,
    config::CONFIG,
    credits::{self, start_credits_process_task},
    db::DbHandle,
    handler,
    jobs::JobHandler,
    processes::{
        BiodiversitySensitiveAreasProcess, ClimateRiskProcess, HabitatDistanceProcess,
        LandUseSealedAreaProcess, NDVIProcess, ProcessesOpenApiSpec,
    },
    state::spawn_with_user,
};
use geoengine_api_client::apis::configuration::Configuration;
use ogcapi::{
    processes::{DynProcessor, echo::Echo},
    services::{self as ogcapi_services},
};
use std::{mem, sync::Arc};
use utoipa::{
    Modify, OpenApi as _,
    openapi::{
        ContactBuilder, OpenApi, Ref,
        schema::{KnownFormat, ObjectBuilder, OneOfBuilder, Schema, SchemaFormat, Type},
    },
};
use utoipa_axum::{router::OpenApiRouter, routes};

#[derive(Clone, Debug)]
pub struct AppState {
    pub db: crate::db::DbHandle,
    pub api_config: Configuration,
}

/// Create and configure the OGC API service, including routes, state, and OpenAPI documentation.
pub async fn server() -> anyhow::Result<ogcapi_services::Service> {
    let db_pool = DbHandle::from_config(&CONFIG.database).await?;

    let mut misc_router = OpenApiRouter::new()
        .routes(routes!(handler::health_handler))
        .nest("/auth", handler::auth_router())
        .nest("/credits", credits::router())
        .merge(handler::profile_router())
        .with_state(AppState {
            db: db_pool.clone(),
            api_config: CONFIG.geoengine.api_config(None),
        });

    misc_router
        .get_openapi_mut()
        .merge(ProcessesOpenApiSpec::openapi());

let mut processors: Vec<Arc<dyn DynProcessor>> = vec![
        Arc::new(Echo),
        Arc::new(NDVIProcess::new(db_pool.clone())),
        Arc::new(ClimateRiskProcess::new(db_pool.clone())),
        Arc::new(LandUseSealedAreaProcess::new(db_pool.clone())),
    ];
    add_habitat_distance_process(&mut processors, db_pool.clone()).await;
    add_biodiversity_sensitive_areas_process(&mut processors, db_pool.clone()).await;

    let drivers = ogcapi_services::Drivers {
        jobs: Box::new(JobHandler::new(db_pool.clone()).await?),
        collections: Box::new(NoCollectionTransactions),
    };

    let ogcapi_state = ogcapi_services::AppState::new(drivers)
        .await
        .processors(processors)
        .sync_process_calls_are_jobs(true)
        .with_spawn_fn(spawn_with_user);

    let mut server_cfg: ogcapi_services::Config = (&CONFIG.server).into();
    if cfg!(test) {
        // Use ephemeral port in tests to avoid "address already in use" when creating the service
        server_cfg.port = 0;
    }

    let mut service = ogcapi_services::Service::try_new(&server_cfg, ogcapi_state)
        .await?
        .processes_api();

    let router = service.get_router_mut();
    *router = mem::take(router)
        .merge(misc_router)
        .layer(GeoEngineAuthMiddlewareLayer);
    add_openapi_info(router.get_openapi_mut());

    let _lookup_join_handle = start_credits_process_task(db_pool.clone());

    Ok(service)
}

/// Modify the OpenAPI spec to add general information about the API, such as title, version, description, contact, etc.
fn add_openapi_info(openapi: &mut OpenApi) {
    ResultsSchemaModifier.modify(openapi);

    openapi.info = utoipa::openapi::InfoBuilder::new()
        .title("BioIS API")
        .version(env!("CARGO_PKG_VERSION"))
        .description(Some("API for the BioIS service, providing access to geospatial processing and job management."))
        .contact(Some(
            ContactBuilder::new()
                .name(Some("Geo Engine GmbH"))
                .url(Some("https://www.geoengine.de/en/"))
                .email(Some("info@geoengine.de"))
                .build()
        ))
        // .terms_of_service(None) // TODO: add link to ToS
        .license(None) // TODO: add link to license
        .build();
}

struct ResultsSchemaModifier;

impl Modify for ResultsSchemaModifier {
    fn modify(&self, openapi: &mut OpenApi) {
        let Some(components) = openapi.components.as_mut() else {
            tracing::warn!("OpenAPI components missing; skipping typed Results schema override");
            return;
        };

        let binary_schema = Schema::Object(
            ObjectBuilder::new()
                .schema_type(Type::String)
                .format(Some(SchemaFormat::KnownFormat(KnownFormat::Binary)))
                .build(),
        );

        let results_schema = OneOfBuilder::new()
            .item(Ref::from_schema_name("NDVIProcessOutputs"))
            .item(Ref::from_schema_name("HabitatDistanceProcessOutputs"))
            .item(binary_schema)
            .title(Some("ExecuteResults"));

        components
            .schemas
            .insert("Results".to_string(), results_schema.into());
    }
}

async fn add_habitat_distance_process(
    processors: &mut Vec<Arc<dyn DynProcessor>>,
    db_pool: crate::db::DbHandle,
) {
    match HabitatDistanceProcess::new(db_pool, "Natura2000").await {
        Ok(habitat_distance_process) => {
            tracing::info!(
                "Successfully initialized HabitatDistanceProcess, adding it to the list of available processes."
            );
            processors.push(Arc::new(habitat_distance_process));
        }
        Err(err) => {
            tracing::warn!(
                "Failed to initialize HabitatDistanceProcess, skipping it: {:#}",
                err
            );
        }
    }
}

async fn add_biodiversity_sensitive_areas_process(
    processors: &mut Vec<Arc<dyn DynProcessor>>,
    db_pool: crate::db::DbHandle,
) {
    match BiodiversitySensitiveAreasProcess::new(db_pool, "Natura2000").await {
        Ok(biodiversity_sensitive_areas_process) => {
            tracing::info!(
                "Successfully initialized BiodiversitySensitiveAreasProcess, adding it to the list of available processes."
            );
            processors.push(Arc::new(biodiversity_sensitive_areas_process));
        }
        Err(err) => {
            tracing::warn!(
                "Failed to initialize BiodiversitySensitiveAreasProcess, skipping it: {:#}",
                err
            );
        }
    }
}