# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BioIS API",
    "description": "API for the BioIS service, providing access to geospatial processing and job management.",
    "contact": {
      "name": "Geo Engine GmbH",
      "url": "https://www.geoengine.de/en/",
      "email": "info@geoengine.de"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/processes": {
      "get": {
        "tags": [
          "Processes"
        ],
        "summary": "Retrieve the list of available processes",
        "description": "The list of processes contains a summary of each process the OGC API - Processes\noffers, including the link to a more detailed description of the process.\n\nFor more information, see [Section 7.9](https://docs.ogc.org/is/18-062/18-062.html#sc_process_list).",
        "operationId": "processes",
        "responses": {
          "200": {
            "description": "Information about the available processe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessList"
                }
              }
            }
          },
          "500": {
            "description": "A server error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                },
                "example": {
                  "status": 500,
                  "type": "https://httpwg.org/specs/rfc7231.html#status.500"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/{jobId}/results": {
      "get": {
        "tags": [
          "Processes"
        ],
        "summary": "Retrieve the result(s) of a job",
        "description": "Lists available results of a job. In case of a failure, lists exceptions instead.\n\nFor more information, see [Section 7.13](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_retrieve_job_results).\n\n",
        "operationId": "results",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The results of a job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Results"
                }
              }
            }
          },
          "404": {
            "description": "The requested URI was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                },
                "example": {
                  "status": 404,
                  "type": "https://httpwg.org/specs/rfc7231.html#status.404"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdditionalParameter": {
        "type": "object",
        "required": [
          "name",
          "value"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "array",
            "items": {}
          }
        }
      },
      "AdditionalParameters": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Metadata"
          },
          {
            "type": "object",
            "required": [
              "parameters"
            ],
            "properties": {
              "parameters": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AdditionalParameter"
                }
              }
            }
          }
        ]
      },
...
      "DescriptionType": {
        "type": "object",
        "description": "Basic description type",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Metadata"
            }
          },
          "additionalParameters": {
            "$ref": "#/components/schemas/AdditionalParameters"
          }
        }
      },
      "Exception": {
        "type": "object",
        "description": "Exception based on [`RFC 7807`](https://datatracker.ietf.org/doc/html/rfc7807)",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "type": "string",
            "description": "A URI reference that identifies the problem type."
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem type."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status code generated by the origin server for this occurrence\nof the problem.",
            "minimum": 0
          },
          "detail": {
            "type": "string",
            "description": "A human-readable explanation specific to this occurrence of the problem."
          },
          "instance": {
            "type": "string",
            "description": "A URI reference that identifies the specific occurrence of the problem.\nIt may or may not yield further information if dereferenced."
          }
        },
        "additionalProperties": {}
      },
...
      "GeoJsonInputMediaType": {
        "type": "string",
        "enum": [
          "application/geo+json"
        ]
      },
...
      "HabitatDistanceProcessOutputs": {
        "type": "object",
        "properties": {
          "habitatCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "habitatName": {
            "type": [
              "string",
              "null"
            ]
          },
          "distanceM": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
      "JobControlOptions": {
        "type": "string",
        "enum": [
          "sync-execute",
          "async-execute",
          "dismiss"
        ]
      },
...
      "Link": {
        "type": "object",
        "description": "Hyperlink to enable Hypermedia Access",
        "required": [
          "href",
          "rel"
        ],
        "properties": {
          "href": {
            "type": "string",
            "description": "Supplies the URI to a remote resource (or resource fragment)."
          },
          "rel": {
            "type": "string",
            "description": "The type or semantics of the relation."
          },
          "type": {
            "type": "string",
            "description": "A hint indicating what the media type of the result of dereferencing\nthe link should be."
          },
          "templated": {
            "type": "boolean",
            "description": "This flag set to true if the link is a URL template."
          },
          "varBase": {
            "type": "string",
            "description": "A base path to retrieve semantic information about the variables used in URL template."
          },
          "hreflang": {
            "type": "string",
            "description": "A hint indicating what the language of the result of dereferencing the\nlink should be."
          },
          "title": {
            "type": "string",
            "description": "Used to label the destination of a link such that it can be used as a\nhuman-readable identifier."
          },
          "length": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Metadata": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "href": {
            "type": "string"
          }
        }
      },
      "Month": {
        "type": "integer",
        "format": "int32",
        "minimum": 0
      },
      "NDVIProcessInputs": {
        "type": "object",
        "required": [
          "coordinate",
          "year",
          "month"
        ],
        "properties": {
          "coordinate": {
            "$ref": "#/components/schemas/PointGeoJsonInput"
          },
          "year": {
            "$ref": "#/components/schemas/Year"
          },
          "month": {
            "$ref": "#/components/schemas/Month"
          }
        }
      },
      "NDVIProcessOutputs": {
        "type": "object",
        "properties": {
          "ndvi": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "kNdvi": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "inputs": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NDVIProcessInputs"
              }
            ]
          }
        }
      },
...
      "PointGeoJsonInput": {
        "type": "object",
        "required": [
          "value",
          "mediaType"
        ],
        "properties": {
          "value": {
            "$ref": "https://geojson.org/schema/Point.json"
          },
          "mediaType": {
            "$ref": "#/components/schemas/GeoJsonInputMediaType"
          }
        }
      },
...
      "ProcessList": {
        "type": "object",
        "description": "Information about the available processes",
        "required": [
          "processes",
          "links"
        ],
        "properties": {
          "processes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProcessSummary"
            }
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            }
          }
        }
      },
      "ProcessSummary": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DescriptionType"
          },
          {
            "type": "object",
            "required": [
              "id",
              "version"
            ],
            "properties": {
              "id": {
                "type": "string"
              },
              "version": {
                "type": "string"
              },
              "jobControlOptions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/JobControlOptions"
                }
              },
              "outputTransmission": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TransmissionMode"
                }
              },
              "links": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Link"
                }
              }
            }
          }
        ],
        "description": "Process summary"
      },
...
      "Results": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/NDVIProcessOutputs"
          },
          {
            "$ref": "#/components/schemas/HabitatDistanceProcessOutputs"
          },
          {
            "type": "string",
            "format": "binary"
          }
        ],
        "title": "ExecuteResults"
      },
...
      "TransmissionMode": {
        "type": "string",
        "default": "value",
        "enum": [
          "value",
          "reference"
        ]
      },
...
      "Year": {
        "type": "integer",
        "format": "int32",
        "description": "Year of reporting or change (e.g., 2023, 2024, etc.)",
        "minimum": 0
      },
...
    }
  }
}